fix: fixed bug with winnerdata constantly playing item break sound

+++ src/client/java/com/roulettepaymenttracker/client/PaymentDataManager.java
@@ -47,23 +47,7 @@ public class PaymentDataManager {
                 System.out.println("Failed to create directories for paymentData.json file: " + exception.getMessage());
             }
 
-            if (Files.exists(paymentDataPath)) { // check if .json file already exists
-                try (Reader fileReader = Files.newBufferedReader(paymentDataPath)) { // reads existing data
-                    Type listType = new TypeToken<List<PlayerDataHolder>>(){}.getType(); // defines expected type od List<PlayerDataHolder>
-                    listOfPlayerData = gson.fromJson(fileReader, listType); // deserialize JSON array into list of PlayerDataHolder objects
-
-                    // if file was null, initialize an empty list to avoid NullPointerException
-                    if (listOfPlayerData == null) {
-                        listOfPlayerData = new ArrayList<>();
-                    }
-                }
-                catch (IOException exception) {
-                    System.out.println("Something went wrong reading existing data: " + exception.getMessage());
-                    actionBarNotification.sendMessage("Can't read data from JSON file.", "§4");
-                    playSoundEffect.playSound(SoundEvents.ENTITY_ITEM_BREAK);
-                }
-            }
-            else {
+            if (!Files.exists(paymentDataPath)) { // check if .json file already exists
                 System.out.println("Couldn't find paymentData.json file.");
                 try {
                     System.out.println("Creating an empty paymentData.json file.");
@@ -78,36 +62,53 @@ public class PaymentDataManager {
                 }
             }
 
-            boolean playerAlreadyExists = false;
-            for (int index = 0; index < listOfPlayerData.size(); index++) {
-                PlayerDataHolder player = listOfPlayerData.get(index);
+            if (Files.exists(paymentDataPath)) {
+                try (Reader fileReader = Files.newBufferedReader(paymentDataPath)) { // reads existing data
+                    Type listType = new TypeToken<List<PlayerDataHolder>>(){}.getType(); // defines expected type od List<PlayerDataHolder>
+                    listOfPlayerData = gson.fromJson(fileReader, listType); // deserialize JSON array into list of PlayerDataHolder objects
 
-                if (player.username().equals(paymentUsername)) {
-                    int updatedAmount = player.amount() + paymentAmount; // updates payment amount
-                    PlayerDataHolder updatedPlayer = new PlayerDataHolder(paymentUsername, updatedAmount); // creates object with updates payment data
-
-                    listOfPlayerData.set(index, updatedPlayer); // puts updated player data in the place of the old data
-
-                    playerAlreadyExists = true;
-                    break; // exits loop
+                    // if file was null, initialize an empty list to avoid NullPointerException
+                    if (listOfPlayerData == null) {
+                        listOfPlayerData = new ArrayList<>();
+                    }
                 }
-            }
+                catch (IOException exception) {
+                    System.out.println("Something went wrong reading existing data: " + exception.getMessage());
+                    actionBarNotification.sendMessage("Can't read data from JSON file.", "§4");
+                    playSoundEffect.playSound(SoundEvents.ENTITY_ITEM_BREAK);
+                }
 
-            if (!playerAlreadyExists) {
-                listOfPlayerData.add(newPlayerData); // adds new player data to list of player data
-            }
+                boolean playerAlreadyExists = false;
+                for (int index = 0; index < listOfPlayerData.size(); index++) {
+                    PlayerDataHolder player = listOfPlayerData.get(index);
 
-            // write the updated list back to the JSON file
-            try (FileWriter fileWriter = new FileWriter(filePath)) {
-                gson.toJson(listOfPlayerData, fileWriter);
+                    if (player.username().equals(paymentUsername)) {
+                        int updatedAmount = player.amount() + paymentAmount; // updates payment amount
+                        PlayerDataHolder updatedPlayer = new PlayerDataHolder(paymentUsername, updatedAmount); // creates object with updates payment data
 
-                System.out.println("Succesfully saved payment data to JSON file");
-                actionBarNotification.sendMessage("Saved payment data to JSON file.", "§a");
-                playSoundEffect.playSound(SoundEvents.ENTITY_VILLAGER_WORK_CARTOGRAPHER);
-            } catch (IOException exception) {
-                System.out.println("Something went wrong during saving data to JSON file: " + exception.getMessage());
-                actionBarNotification.sendMessage("Couldn't save payment data to JSON file.", "§4");
-                playSoundEffect.playSound(SoundEvents.ENTITY_ITEM_BREAK);
+                        listOfPlayerData.set(index, updatedPlayer); // puts updated player data in the place of the old data
+
+                        playerAlreadyExists = true;
+                        break; // exits loop
+                    }
+                }
+
+                if (!playerAlreadyExists) {
+                    listOfPlayerData.add(newPlayerData); // adds new player data to list of player data
+                }
+
+                // write the updated list back to the JSON file
+                try (FileWriter fileWriter = new FileWriter(filePath)) {
+                    gson.toJson(listOfPlayerData, fileWriter);
+
+                    System.out.println("Succesfully saved payment data to JSON file");
+                    actionBarNotification.sendMessage("Saved payment data to JSON file.", "§a");
+                    playSoundEffect.playSound(SoundEvents.ENTITY_VILLAGER_WORK_CARTOGRAPHER);
+                } catch (IOException exception) {
+                    System.out.println("Something went wrong during saving data to JSON file: " + exception.getMessage());
+                    actionBarNotification.sendMessage("Couldn't save payment data to JSON file.", "§4");
+                    playSoundEffect.playSound(SoundEvents.ENTITY_ITEM_BREAK);
+                }
             }
         }, executorService); // makes the method use dedicated thread pool for execution
     }
